namozvaqtlari/requirements.txt
django
django-jazzmin
django-ninja
pytelegrambotapi
django-kronos
beautifulsoup4